update grenade info
